+++ R/survey_to_diffnet.R
@@ -112,7 +112,7 @@ check_var_class_and_coerce <- function(var, dat, class.ok, class.target, warn.co
 #'    fakesurveyDyn, "id", c("net1", "net2", "net3"), "toa", "group" ,
 #'    timevar = "time", keep.isolates = TRUE, warn.coercion=FALSE)
 #'
-#' plot_diffnet(x, vertex.cex = 1.5, label = rownames(x))
+#' plot_diffnet(x, vertex.label = rownames(x))
 #'
 #' # Reproducing medInnovationsDiffNet object ----------------------------------
 #' data(medInnovations)